refactoring Windows structures
