+++ Windows/PerfOptions.py
@@ -1,6 +1,6 @@
 from PyQt5.QtWidgets import QApplication, QWidget, QComboBox, QGroupBox, \
                 QVBoxLayout, QRadioButton, QLabel, QSlider, QPushButton, QMessageBox
-from Windows.PerfOptionsTemplate import Ui_Dialog
+from Windows.Templates.PerfOptionsTemplate import Ui_Dialog
 from Windows.SimplePerfOptions import SimplePerfOptions
 from PopUp import PopUpWrapper
 import re
