--- Windows/widget_pane.json
@@ -1,86 +1,0 @@
-{
-    "CubicGLContext": {
-                      "alias": "3D_CUBIC",
-                      "object": "CubicGLContext",
-                      "object_type": "3d_object",
-                      "settings": ["PerfOptions", "file_header",
-                                   "object_alias"],
-                      "required": ["current_state", "options",
-                                   "file_header", "color_vectors",
-                                   "screenshot_dir", "geom"],
-                      "optional": ["iterations"],
-                      "iterable_type": "structure",
-                      "toolbar" : [["Reset", "initial_transformation"],
-                                   ["Zoom in", "zoomIn"],
-                                   ["Zoom out", "zoomOut"],
-                                   ["Screenshot", "screenshot_manager"],
-                                   ["Text", "popup_text_selector"]]
-                      },
-    "VectorGLContext": {
-                      "alias": "3D_VECTOR",
-                      "object": "VectorGLContext",
-                      "object_type": "3d_object",
-                      "settings": ["PerfOptions", "file_header"],
-                      "required": ["current_state", "options",
-                                   "file_header","color_vectors",
-                                   "screenshot_dir", "geom"],
-                      "optional": ["iterations"],
-                      "iterable_type": "structure",
-                      "toolbar" : [["Reset", "initial_transformation"],
-                                   ["Zoom in", "zoomIn"],
-                                   ["Zoom out", "zoomOut"],
-                                   ["Screenshot", "screenshot_manager"],
-                                   ["Text", "popup_text_selector"]]
-                      },
-
-    "ArrowGLContext": {
-                      "alias": "3D_ARROW",
-                      "object": "ArrowGLContext",
-                      "object_type": "3d_object",
-                      "settings": ["ArrowPerfOptions", "file_header", 
-                                   "object_alias"],
-                      "required": ["current_state", "options",
-                                   "file_header","color_vectors",
-                                   "screenshot_dir", "geom"],
-                      "optional": ["iterations"],
-                      "iterable_type": "structure",
-                      "toolbar" : [["Reset", "initial_transformation"],
-                                                   ["Zoom in", "zoomIn"],
-                                                   ["Zoom out", "zoomOut"],
-                                                   ["Screenshot", "screenshot_manager"],
-                                                   ["Text", "popup_text_selector"]]
-                      },
-    "Canvas":         {
-                        "alias": "2D_Matplotlib",
-                        "object": "Canvas",
-                        "object_type": "2d_object",
-                        "settings": ["PlotSettings", "plot_data"],
-                        "required": ["iterations", "current_state", "options",
-                                     "plot_data",  "screenshot_dir"],
-                        "optional": ["trigger"],
-                        "iterable_type": "plot",
-                        "toolbar": "NavigationToolbar"
-                        },
-    "Canvas2Dupgraded": {
-                        "alias": "2D_BetterPlot",
-                        "object": "Canvas2Dupgraded",
-                        "object_type": "2d_object",
-                        "settings": ["PlotSettings", "plot_data"],
-                        "required": ["iterations", "current_state", "options",
-                                     "plot_data", "geom", "screenshot_dir"],
-                        "optional": ["trigger"],
-                        "iterable_type": "plot"
-                      },
-    "CanvasLayer": {
-                        "alias": "2D_Layer",
-                        "object": "CanvasLayer",
-                        "object_type": "2d_object",
-                        "settings": ["SimplePerfOptions", "file_header"],
-                        "required": [ "current_state", "options",
-                                     "file_header","color_vectors",
-                                     "screenshot_dir"],
-                        "optional" : ["iterations"],
-                        "iterable_type": "structure",
-                        "toolbar": "NavigationToolbar"
-                      }
-}
