+++ makefile
@@ -1,10 +1,10 @@
 convert:
-	pyuic5 -x Windows/UI_files/MainWindow.ui -o Windows/MainWindowTemplate.py
-	pyuic5 -x Windows/UI_files/animationSettings.ui -o Windows/animationSettingsTemplate.py
-	pyuic5 -x Windows/UI_files/PlotSettings.ui -o Windows/PlotSettingsTemplate.py
-	pyuic5 -x Windows/UI_files/PerfOptions.ui -o Windows/PerfOptionsTemplate.py
-	pyuic5 -x Windows/UI_files/SimplePerfOptions.ui -o Windows/SimplePerfOptionsTemplate.py
-	pyuic5 -x Windows/UI_files/Select.ui -o Windows/SelectTemplate.py
+	pyuic5 -x Windows/UI_files/MainWindow.ui -o Windows/Templates/MainWindowTemplate.py
+	pyuic5 -x Windows/UI_files/animationSettings.ui -o Windows/Templates/animationSettingsTemplate.py
+	pyuic5 -x Windows/UI_files/PlotSettings.ui -o Windows/Templates/PlotSettingsTemplate.py
+	pyuic5 -x Windows/UI_files/PerfOptions.ui -o Windows/Templates/PerfOptionsTemplate.py
+	pyuic5 -x Windows/UI_files/SimplePerfOptions.ui -o Windows/Templates/SimplePerfOptionsTemplate.py
+	pyuic5 -x Windows/UI_files/Select.ui -o Windows/Templates/SelectTemplate.py
 
 cython:
 	python3 cython_modules/ex_setup.py build_ext --build-lib cython_modules/ --inplace
